Factored out cli into cli.py
