+++ clonepool/__main__.py
@@ -5,7 +5,7 @@ https://kushaldas.in/posts/building-command-line-tools-in-python-with-click.html
 '''
 import click
 
-from clonepool.clonepool import layout_cli, simulate_cli, resolve_cli
+from clonepool.cli import layout_cli, simulate_cli, resolve_cli
 
 # Do not sort the command list alphabetically, but use order in which commands
 # are added.
