+++ clonepool/clonepool.py
@@ -3,9 +3,7 @@
 from collections import defaultdict
 import sys
 
-import click
 import numpy as np
-
 from clonepool.utils import (
     set_up_pools,
     simulate_pools,
@@ -16,30 +14,6 @@ from clonepool.utils import (
     read_layout_file,
     write_layout_file,
 )
-
-
-@click.command('layout')
-@click.option(
-    '-n', '--samples', required=False, type=int,
-    help='Number of samples')
-@click.option(
-    '-P', '--pool-size', required=True, type=int,
-    help='How many samples go into each pool')
-@click.option(
-    '-r', '--replicates', default=2, type=int,
-    help='Number of replicates per sample [2]')
-@click.option(
-    '-w', '--pool-count', default=94, type=int,
-    help='Number of pools (wells) [94]')
-@click.argument(
-    'layout_file', required=False, default='-', type=click.File('w'))
-def layout_cli(pool_size, pool_count, replicates, samples, layout_file):
-    '''
-    Generate pool layout. This assigns all samples to their respecitve pools.
-
-    Writes to STDOUT or the given layout file.
-    '''
-    layout(pool_size, pool_count, replicates, samples, layout_file)
 
 
 def layout(pool_size, pool_count, replicates, samples, layout_file):
@@ -57,32 +31,6 @@ def layout(pool_size, pool_count, replicates, samples, layout_file):
     write_layout_file(layout_file, pool_log)
 
 
-@click.command('simulate')
-@click.option(
-    '-l', '--layout', required=True, type=click.File('r'),
-    help='Path to input file containing pool layout')
-@click.option(
-    '-p', '--prevalence', default=0.05, type=click.FloatRange(0, 1),
-    help='Sample prevalence used for simulation [0.05]')
-@click.option(
-    '-P', '--false-positives', default=0, type=click.FloatRange(0, 1),
-    help='Fraction of false-positive pools [0]')
-@click.option(
-    '-N', '--false-negatives', default=0, type=click.FloatRange(0, 1),
-    help='Fraction of false-negative pools [0]')
-@click.argument(
-    'out_layout_file', required=False, default='-', type=click.File('w'))
-def simulate_cli(layout, prevalence, false_positives, false_negatives, out_layout_file):
-    '''
-    For a given pool layout, simulate a test run. Uses a defined sample
-    prevalence to determine a random set of positive samples and,
-    successively, flags all pools as positive containing any of these samples.
-
-    Writes to STDOUT or the given layout file.
-    '''
-    simulate(layout, prevalence, false_positives, false_negatives, out_layout_file)
-
-
 def simulate(layout, prevalence, false_positives, false_negatives, out_layout_file):
     # Read existing pool layout, discard old positive pools / samples if any.
     pool_log, _, _ = read_layout_file(layout)
@@ -98,21 +46,6 @@ def simulate(layout, prevalence, false_positives, false_negatives, out_layout_fi
     # Write layout including new pool results.
     write_layout_file(
             out_layout_file, pool_log, positive_pools, positive_samples)
-
-
-@click.command('resolve')
-@click.option(
-    '--layout', '-l', required=True, type=click.File('r'),
-    help='Path to layout file containing +/- pool results')
-@click.argument(
-    'sample_results_file', required=False, default='-', type=click.File('w'))
-def resolve_cli(layout, sample_results_file):
-    '''
-    Resolve sample status from pool results. As this is not always
-    possible, some samples may remain in an uncertain state.
-    Writes to STDOUT or the given results file.
-    '''
-    resolve(layout, sample_results_file)
 
 
 def resolve(layout, sample_results_file):
@@ -138,5 +71,3 @@ def resolve(layout, sample_results_file):
     for sample, state in sorted(sample_state.items()):
         state_symbol = '+' if state == +1 else '-' if state == -1 else 'NA'
         sample_results_file.write(f'{sample}\t{state_symbol}\n')
-
-# EOF
